fix pow() ambiguity for VC++

+++ pix_opencv_of_hs.cc
@@ -140,7 +140,7 @@ void pix_opencv_of_hs :: processRGBAImage(imageStruct &image)
         dest.x = (int)(orig.x + cvGet2D(x_velx, py, px).val[0]);
         dest.y = (int)(orig.y + cvGet2D(x_vely, py, px).val[0]);
         angle = -atan2( (double) (dest.y-orig.y), (double) (dest.x-orig.x) );
-        hypotenuse = sqrt( pow(dest.y-orig.y, 2) + pow(dest.x-orig.x, 2) );
+        hypotenuse = sqrt( pow((double)dest.y-orig.y, 2) + pow((double)dest.x-orig.x, 2) );
 
         /* Now draw the tips of the arrow. I do some scaling so that the
         * tips look proportional to the main line of the arrow.
@@ -268,7 +268,7 @@ void pix_opencv_of_hs :: processRGBImage(imageStruct &image)
         dest.x = (int)(orig.x + cvGet2D(x_velx, py, px).val[0]);
         dest.y = (int)(orig.y + cvGet2D(x_vely, py, px).val[0]);
         angle = -atan2( (double) (dest.y-orig.y), (double) (dest.x-orig.x) );
-        hypotenuse = sqrt( pow(dest.y-orig.y, 2) + pow(dest.x-orig.x, 2) );
+        hypotenuse = sqrt( pow((double)dest.y-orig.y, 2) + pow((double)dest.x-orig.x, 2) );
 
         /* Now draw the tips of the arrow. I do some scaling so that the
         * tips look proportional to the main line of the arrow.
@@ -396,7 +396,7 @@ void pix_opencv_of_hs :: processGrayImage(imageStruct &image)
         dest.x = (int)(orig.x + cvGet2D(x_velx, py, px).val[0]);
         dest.y = (int)(orig.y + cvGet2D(x_vely, py, px).val[0]);
         angle = -atan2( (double) (dest.y-orig.y), (double) (dest.x-orig.x) );
-        hypotenuse = sqrt( pow(dest.y-orig.y, 2) + pow(dest.x-orig.x, 2) );
+        hypotenuse = sqrt( pow((double)dest.y-orig.y, 2) + pow((double)dest.x-orig.x, 2) );
 
         /* Now draw the tips of the arrow. I do some scaling so that the
         * tips look proportional to the main line of the arrow.
